chore: exclude useless GPU type in resource pool API [DET-5079] (#2016)
In GCP, when the number of GPUs is zero, do not include the GPU type as part of the instance type string. This avoids confusing strings such as "n1-standard-4, 0xnvidia-tesla-k80".

Also adds a space between the number of GPUs and the type so that it is easier to read.