feat: add trial profiler metrics APIs [DET-5065, DET-5059] (#2051)
This change adds implementations for the trial profiler metrics to be added by the observability project. This includes:
  * GetTrialProfilerMetrics API and tests
  * GetTrialProfilerAvailableSeries API and tests
  * PostTrialProfilerMetricsBatch API and tests
  * trial_profiler_metrics table and trial_profiler_metric_labels index on the table.
All of this work is discussed in depth in the ERD for the observability project. These changes mimic trial logs design, and don't preclude an escape hatch to ship to a real TSDB for large installs, though for the most part these features are for small and ephemeral profiling data just to aid users in porting, and it won't be large likely.