feat: add allocation aggregation by agent label and resource pool
Just add some new things parallel to the existing ones, pretty much.

A few existing field names are edited to be clearer, by request. That
will break behavior on previously computed aggregations, and also the
new aggregations won't be filled in for any databases that have already
been run with the previous version, but I consider all that acceptable
because we haven't had a release since that landed. If we ever need to
add more aggregations in the future, we'll figure it out then.

Also disable a Go lint that was raising an error on some of the new code
but turns out to be obsolete (and replaced by other lints that are still
enabled and not complaining) according to its documentation.