chore: clarify API for expconf objects
Make the API clear that the getters and setters are the right way to
work with expconf objects.  Rename all fields to be RawDescription,
RawHyperparameters, etc.

The RawField might be a pointer, but its getter, Field(), will only be a
pointer if no default exists for schemas.WithDefaults() to fill in.

For operations which need to manipulate the expconf object before
schemas.WithDefaults() is called, the RawField may be read or set
directly.

The RawField must be a public field, since the reflect library can't
work with private fields at all(!).