fix: gracefully handle Docker binding published ports to ipv4 and ipv6 for host (#2259) [DET-5295]
This change updates how we determine the Addresses for a container to gracefully handle when Docker publishes ports for ipv4 and ipv6 (e.g. looks like `0.0.0.0:8080->8080/tcp, :::8080->8080/tcp`) by coalescing either to `agent-ip:host-port:...` and de-duplicating the Addresses returned to consumers when ipv4 and ipv6 are coalesced to the same address.