chore: fix common fields on union types (#2270)
It is way too complex to allow unions to contain common fields, which
can be specified by themselves (without any union member) and merged
into other fully-specified structures.  This was, in fact, exactly how
the old union marshaling and unmarshaling worked.

This complexity is why the master yaml has a checkpoint storage config
which is just a slice of bytes; it is possible to specify partial values
which cannot be represented as a CheckpointStorage object at all; they
can only be unmarshaled into other objects.

The json-schema definitions originally tried to eliminate this
complexity by doing away with common fields altogether and putting
duplicate fields on the union member structs.  That way the typing is
fairly simple to understand; in python terms:

    CheckpointStorage = Union[
        SharedFSConfig, S3Config, GCSConfig, HDFSConfig
    ]

However, that does not support the common-field-merging behavior or the
partial-config-specification behavior that we have explicitly offered
for CheckpointStorage in the master config, and also for templates.

This change redefines the v0 json schema for checkpoint storage,
searcher, and data layer to support to more accurately reflect the
current behavior, for hopefully a seamless user experience.

Interestingly, hyperparameters in the old system did not support partial
definitions like the other union types, due to the behavior of json
unmarshaling into maps.

Future versions of the expconf schema will have simpler mechanics.