chore: loosen ruamel.yaml version
Users are seeing messages about the wrong ruamel version being
installed.

I'm not sure which package is resulting in this old verison, but it
doesn't much matter; the reason we required 0.15.78 was due to our old
runtime image building system, where we had too much control over the
versions that got installed.  Now, we support ruamel.yaml as far back as
0.15.29.