chore: add zmq-based IPC to the DistributedContext
This is part of a multi-step process to refactor the harness code in
favor of push architecture:

Goal:

    Introduce Python WorkloadSequencer that uses Push Architecture.

That can't happen until:

    The WorkloadManager is broken into simple function calls for
    mutating workload responses.  These simple functions (one after each
    of training, validation, and checkpointing) can be called right
    before making calls to the push API, rather than after responding
    with a workload.Response on a stream of workloads.

That can't happen until:

    The StorageManager is no longer a part of the WorkloadManager.  The
    StorageManager is the only part of the WorkloadManager that doesn't
    fit neatly into the aforementioned three function calls, since
    storage_mgr.store_path() is a contextmanager.  However, if the
    TrialController were responsible for the StorageManager rather than
    the WorkloadManager, the extra `args: List[Any]` parameter in the
    workload stream would disappear.

Now, it would be nice to package all of that functionality on a
GenericContext of some sort, in preparation for the eventual Generic
API.  It makes sense that the GenericContext would be the place that all
metrics and checkpoint reporting would go through, so that would be the
place where after_training(), after_validation(), and
after_checkpointing() would each be called.  It also makes sense that
the GenericContext would expose the StorageManager to user code.

But nothing can be packaged into the GenericContext until the
GenericContext has ZMQ communications.  For example, a
GenericContext.maybe_load_checkpoint() can't function without
cross-worker synchronization; only the local_rank=0 worker each compute
node should actually load the checkpoint, and the other workers must
wait.

We have all of the tools for ZMQ communication in the harness, and we
even have the ports allocated for it.  But they have always been
attached to the TrialController, rather than the DistributedContext.
That is because right after the refactor, I believed/enforced that the
Context objects were purely for light-weight API call exposures and
didn't have any beefy logic in them.  However, PyTorchContext has proven
that putting logic in the Context object is actually a great strategy,
and the upcoming Generic API is really just the logical conclusion of
that realization.

So as a first step, the ownership of the ZMQ sockets is passing to the
DistributedContext (which will eventually be a member of the
GenericContext).