chore: close agents on websocket closures (#2380)
This change updates the agent tracking code in the master to terminate when the websocket actor closes. Prior to #2351 it seems the agent actor in the master relied on the websocket actor propogating actor.ChildFailed up to it to detect exits and mark agents as terminated. #2351 changed this to swallow abnormal websocket closures.