fix add schemas.Defaultable and schemas.Copyable
The reflect-based behavior of schemas.WithDefaults and schemas.Merge
works great on plain-old-data types which have no private fields, but it
pukes on private fields.

There already was a schemas.Mergable interface for customizing the
merging behaviors of objects.  This change introduces the
schemas.Defaultable and schemas.Copyable interfaces, which have
analagous effects on schemas.WithDefaults and schemas.cpy (unexported).

schemas.Copyable is very easy to understand.  For example, the new
expconf.Pod (which is type-defined to k8sV1.Pod) simply calls
k8sV1.Pod.DeepCopy().

schemas.RuntimeDefaultable has been removed in favor of
schemas.Defaultable.  Defaultable makes the reflect code a little
easier and is more like the Mergable and Copyable
interfaces than RuntimeDefaultable was.  The flipside is that the
runtime default behavior of the ExperimentConfig.Description is slightly
more complex.  Overall the tradeoff seems appropriate; the general code
(schemas.WithDefaults()) got more general and clearer, and the specific
challenges of the calling code are handled in the calling code.

Overall, the motivation for this change was to protect against panics
when interfacing with the k8sV1.Pod object.  The only other external
object in the expconf project, `AuthConfig`, is reflect-friendly and
does not require modification.