feat: support push metric APIs internally [DET-5215] (#2315)
This change adds push APIs for trials to push training metrics, validation metrics and checkpoint metadata to the master. To remain backwards compatible with the step data model while allowing us to break away from it in the future, start_time is not required on push but derived from the end time of the last workload or start of the current trial run. Additionally, so that push APIs are unable to permanently delete loads of data with misuse, we also moved to storing multiple runs rather than rolling them back as is done currently in the trial actor.