chore: only return the port binding appropriate for the proxy [DET-5495] (#2401)
This change alters how we derive container addresses. When docker changed to make ipv6 show up by default again it threw us for a loop. My hasty fix wasn't quite correct but also introduced with it a bug that also seemed to be docker port related. This lead me to hackily fix the bug of the bug fix rather than realize the root cause, since a revert of the original fix of bug caused the docker ipv6 change to not allow any containers to run. This change fixes both issues by always just returning a single container address when the container is bound to 0.0.0.0 or :: based on whether the proxy is ipv4 or ipv6.