fix: make profiling schema more lenient [DET-5497] (#2409)
Users should be able to set "profiling.enabled: true" in the experiment config and fall back to intelligent defaults on starting batch and duration of profiling. The current experiment config doesn't allow only "enabled: true" because it wants to validate that "begin_on_batch" <= "end_after_batch". This PR changes the experiment config schema to allow one of both of the fields to be unset when enabled: true

This is fine because:
- if both are unset, the default behavior of starting at batch 0 and running for 5 minutes is good
- if start is unset but the end is set, it will default to 0 (previously this was handled by logic in the experiment config, but this PR moves that logic into the schema itself). Then the end doesn't matter because it should always be >= 0 (do we need to add a check to enforce this? it's pretty common sense)
- if the end is unset, but the start is set, the behavior is to profile for 5 minutes which is good behavior.