fix: prevent spot resource pool contention [DET-5349] (#2423)
There is a bug where spot requests and spot instances in AWS do not include the resource pool as part of the tags. This leads to issues where one spot resource pool creates instances, but all of the spot resource pools think they are responsible for managing the instance. This causes spot instances to get shut down because some resource pools think they are idle. This PR ensures that we include the resource pool as one of the tags to prevent this contention.

This PR also ensures that this change doesn't lead to orphaned spot instances by deleting all spot requests and instances that do not include the "determined-resource-pool" tag when the provisioner starts up.