chore: bump buf and protoc version [DET-5534] (#2446)
This change bumps the buf and protoc version. Previously we were locked to buf 0.12.1 which was only aware of google's well know types (wkt's) on github not google.golang.com so we were all good. A buf image generated with 0.42.1 was push which lead to CI using 0.12.1 thinking there was a breaking change since the location of wkt's changed. Since we were 30 minor version behind anyways, this was due.