chore: catch ruamel.yaml Duplicate Key Errors and format for users [DET-5542] (#2450)
If `ruamel.yaml`'s `safe_load` is called directly on a faulty yaml file, it gives the user an error with a traceback including our internal apis. This PR adds a utility method that calls `ruamel.yaml`'s `safe_load`, but also catches any errors it may raise, and formats them neatly for the user.