chore: measure profiler timings with time.time() (#2475)
Change ProfilerAgent to record timings using time.time() instead of datetime.utcnow(). For timing small segments of code, this will make the measurements more accurate as utcnow() does conversion work that was previously included as part of the measurement.

time.time() was chosen over time.time_ns() because that was introduced in Python 3.7 and at the time of coding we needed to support 3.6. It was chosen over time.perf_counter() and time.monotonic() because they are only useful for relative measurements (i.e. calculating durations). We need to record a timestamp to send to the DB so using those clocks would require us to store a timestamp in addition to the starting perf_counter/monotonic value. That seems to complicate the code more than the benefits are worth - we don't think the additional granularity of perf_counter is useful (the webui displays millisecond granularity) and we don't expect the system clock to change during training.

Also change all datetimes to be explicitly UTC so they are timezone aware.