fix: ProfilingAgent serializing timestamps incorrectly (#2482)
Fix bug where some datetimes were being serialized in a valid format, while others were not and causing POSTs to fail.

Python's datetime.datetime objects can be naive or timezone-aware. When isoformat() is called on them, the output is different depending on whether it is timezone-aware. During the initial implementation, all datetimes were naive, which meant the serialized form did not include any timezone information. We appended 'Z' to all timestamp strings to indicate that they are in UTC. PR #2475 changed most of the datetimes to be timezone-aware so appending 'Z' to all timestamps led to "2021-06-01T21:41:23.411254+00:00Z" instead of the intended "2021-06-01T21:41:23.411254Z". Now that we are working with timezone aware datetimes, the best way to serialize them is to just call isoformat() to get "2021-06-01T21:41:23.411254+00:00" which is valid.

However, this exposed the fact that datetime.utcnow() actually creates a datetime that is timezone-unaware, which caused new errors. So all uses of datetime.utcnow() were replaced with the preferred datetime.now(timezone.utc) which is timezone aware. And now the serialization logic performs a check to ensure that all datetimes are timezone-aware and in UTC to avoid this headache in the future.