feat: support .tgz model definitions
The determined-master has always taken an unorthodox
tarball-over-json-like file format for specifying context directories
(notably for POST experiment, but also for all varieties of commands).

The new format is to post the contents of a gzipped tarball directly.

In all cases, the master supports both the old and the new formats via
different arguments.  This is important because it is fairly common for
users to run old cli's against new masters, and it is only a small
amount of shimming code to give those users an excellent experience.