chore: allow non-singleton Authentication
The authentication.Authentication object was written as a singleton
object for the CLI.  Since the CLI never talked to multiple masters or
wrote to multiple masters, that pattern worked great.

However, the upcoming Python SDK will need more flexibility than that.
Rather than duplicate the existing authentication logic, we can refactor
what exists now to be a bit more flexible so that the Python SDK can
reuse what is already written.