chore: improved pynvml usage by profiler [DET-5394] (#2487)
Wrap pynvml in a class that checks if pynvml is available and if pynvml is working. Abstracts away these checks from the user and just exposes a is_available boolean. Also wraps the various pynmvl APIs so that they are now typed and have some basic caching.

Also improves the design of the GPU memory/util collectors so that they will not cause errors if used when there are not GPUs (they just return an empty dict of measurements).