chore: collect samples per second for pytorch and keras [DET-5515] (#2537)
This change adds "misc" metrics and collects samples_per_second as the first one, so we can display it in the webui again, but not rely on getting it from a step's metrics (since those will go away shortly). It also heavily refactors the profiler code to have a single "batcher" type and a single "batcher thread" type, instead of one specifically for timings/system/misc metrics.