chore: break down WorkloadManager [DET-5419]
Remove the determined.callback.Callback, which only had one
implementation in existence.  Just inline those calls.

Reduce the WorkloadManager to a trio of hooks, and attach them to a new
_GenericContext.  Eventually _GenericContext will be cleaned up, exposed
as a public, and will become the basis of the Generic API.

Each TrialController instantiates its own _GenericContext, which is
basically how the TrialControllers will work in the future.

Significantly, this moves all major logic out of the parent processes
and into the the worker processes.  All that remains in the parent
processes is the SocketManager and the calling of horovodrun.  The
SocketManager will be removed after push architecture is complete,
leaving only horovodrun, which will eventually evolve into a
configurable launch layer.