feat: add rendezvous API (#2420) [DET-5428]
This change adds a rendezvous API for trials in a post-push-arch world. The API is designed to block until all containers are connected and their rendezvous info has arrived from container started message. It containers take too long to connect, the API will timeout and print a warning. If not, once the conditions are met, rendezvous is pushed to listening watchers.