fix: switch container runtime on slot type (#2845)
Previous we set the default container runtime to `nvidia` regardless of slot type for the agent. This change moves us to selecting `runc` or `nvidia` based on the slot type, since nvidia-based GPU images won't launch on CPU agents with the `nvidia` runtime when drivers/etc are already baked in (it tries to run an container start hook and fails).