feat: add sync timing toggle for profiling (#2874) [DET-5891]
This change adds the toggle `profiling.sync_timings` which defaults to `true` and forces Determined to call `torch.cuda.synchronize` appropriately so that timing metrics are more accurate. Since this comes as a trade off, the toggle is available, which disables synchronize and instead of collecting many granular timing metrics, the profiler instead collects only throughout and system metrics.