chore: add warnings on resource manager exits (#2903)
This change adds warning logs when the Determined resource manager crashes. Previously the master used the default "rootActor" which ignored all messages. This adds a root actor just for the master that warns on failures of critical sub-components of Determined and, if it returns an error, it has the potential to crash the master (though it doesn't do that quite yet, we need a better recovery strategy first).

Previously, there were incidents where the `pods` actor crash didn't result in a system crash and we never recovered, but it was non-obvious to the user what had happened. The pods actor crashes were addressed in #2880, this is just the next in a line of improvements. In general we should _never_ see this message, but if it happens I want it to at least be obvious.

Nice to have future work after this PR would be to handle cleaning up stray agents/pods on startup (easy) or recovering them properly (hard) and causing RM crashes to crash the master (returning an error from this root actor).