feat: introduce ClusterInfo API (#2946)
This is not yet a public API, but eventually it will be a public API
primarily consumed by custom launch layers.

Users and developers alike should avoid depending on the ClusterInfo API
from within training code.  Code which depends the ClusterInfo API is
code that will never run outside of a Determined cluster.  That's OK for
a launch layer, but not ideal for a training script.  Instead, pass
information into the training script from the launch layer, so that a
drop-in replacement launch layer could be written to run the same
training script on different hardware.