fix: gracefully handle prestart agent failures (#3049)
Fix a bug that could crash the resource manager when an agent disconnects before it sent the master it's AgentStarted message. In this case, the failure caused the master to move the agent to the draining state but, since it wasn't registered with the resource manager, this made the RM panic.