refactor: let shared_fs be "magic"
A while ago, I decided that all forms of "magic" were wrong and that you
should always have to create a storage manager with an explicit
configuration (`container_path: Optional[str]`) that clearly indicated
if you were inside a container or on the host machine.

The idea was to have all the "magic" of detecting where you were
operating in the higher layers of the callstack, and to let the inner
layers of the callstack be as predictable and as explicit as possible.

But you know what?  The resulting API was terrible in all cases.  I
think that correct behavior of `shared_fs` storage manager is magical
by definition.

The result was that there was an overall crappy API, and there was a
missing feature: no super-simple FileStorageManager that took only a
single path as input.  While it's not clear if such a storage manager
would be widely useful to users on the cluster, the unit tests would be
a lot clearer if such a thing existed, and some day local training mode
would be too.

So to solve both of those problems, we let shared_fs be exactly as
magical as it was originally designed (based on the new ClusterInfo API)
and we introduce a FileStorageManager (currently, only accessible in
python).