fix: non-scalar nan handling (#3143)
As a quick release-party fix, add a post-processing step to catch NaN's
in arrays.

This means we do a pre-processing step and a post-processing step. This
is less code (better for a release party fix) but the long-term fix
will involve replacing simplejson.dumps, which doesn't allow us to
customize handling of python-native types.