chore: remove simplejson
simplejson is just externally-maintained json.  We don't have any
particular need to bring in an external dependency over what's packaged
in python's standard library, so just use the built-in json module.

Additionally, stop using json's "default" argument for encoding custom
types.  The function passed to "default" argument is not consulted when
encoding default types (especially "float") so it is impossible to catch
NaNs.  As a result, when using "default", we had to pre-and-post-process
the object during encoding.  Now, we do our own pre-processing step,
where we explicitly visit every object so we can eliminate the
post-processing step.