fix: adhere to proper SemVer format
According to the spec [1], things like `dev0` and `rc0` need to be
separated from the main version numbers by a hyphen rather than a period
or nothing at all. This mostly doesn't really matter, except that we
would like the Helm chart's version to match the Determined version, and
Helm is actually strict about only accepting the hyphen form. So this
changes everything over and updates the current Helm version to match,
which will allow it to be updated by bumpversion from now on.

Awkwardly, wheels actually come out with slightly different modified
versions (per the normalization specified by PEP 440), so we have to add
a little shim to get wheels copied over to task containers successfully.
We also now suppress the warning that `python setup.py build` emits when
the normalized version is different from the original version.

[1] https://semver.org