chore: rename Generic API to Core API [DET-6243]
"Generic" was too... generic.  It doesn't communicate anything.

"Lite" was proposed, but suggests an API that is not full-featured.

"Sprinkle" was proposed, but its name is based too closely on our
history of plugin designs, which won't always be relevant.

"Core" suggests an API that is low-level and also suggests that other
higher-level APIs are likely built on top of it.