fix: remove old build targets (#3268)
* fix: remove test-cli from make test

Tests were moved from ./cli to ./harness in #2108 but the make target
was not updated. Here we remove it so that the absence of the tests
directory does not cause `make test` to fail.

* fix: remove test-common from make test

Tests were moved from ./common to ./harness in #2018 but the make target
was not updated. Here we remove it so that the absence of the tests
directory does not cause `make test` to fail.