fix: generate SSH keys for restored trials (#3249)
Previously, the SSH RSA key size for trials was set on a code path that
was only run for newly created experiments. A trial restored after a
master restart would therefore attempt to generate keys with a size of 0
and crash. Instead, we now set the key size in the master's base task
spec so that it applies to all tasks automatically.