fix: don't allow allocations to take actions with unreceived cancellations (#3326)
A bug occurred where, in quick succession, a RM granted an allocation resources and it reached out to build its task spec/save the trial to the DB, and the trial was canceled. This caused a trial to get saved in the ACTIVE state (as it always defaults to) instead of its actual state STOPPING_CANCELED. Rather than just save the state correctly by saving it as STOPPING_CANCELED, we introduce a new error type and return it to the allocation when it takes an action with an unreceived cancellation in its inbox. This stops the task from pointlessly allocating then killing itself and prevents the bug since we only add an active trial when the trial is active.