ci: print trial logs in more failure cases (#3333)
Previously, if an experiment failed to complete within a timeframe, then
it failed to cancel within a timeframe, the trial logs were never
printed.

There was logic in place to try to avoid printing too many trial logs
when there was a mix of successful and failure trials in the same
experiment.  In practice, the number of e2e tests with multiple trials
is extremely low, but the edge cases where tests failed to show useful
logs was non-zero.  It's much better for developers to just always print
all trial logs when the wait_for_state fails.