fix: handle calls to old command endpoints, make it harder to crash cmd managers [DET-6336] (#3315)
This change updates the command manager code to not crash permanently when receiving API calls from echo routed through the actor system. The actor's handling of these calls was removed but the system still could route calls to it, so, for example, an old CLI could crash commands until a restart occurred. We still need to route to the actors below `/commands/...`, `/shells/...`, etc, for the event endpoints et al, so the routing cannot just be purged.