chore: AdvancedSearcher->Searcher (#3339)
Turns out that the master doesn't even need units when configuring the
searcher.  That means that the basic searcher doesn't make much since,
since it only really worked with a large unit (epochs) rather than a
small unit (batches/records), so for now the Core API will only support
SearcherOp-based training loops.

This is a bit unforunate, since a SearcherOp-based training loop
requires an extra layer of for loop that is not totally intuitive.  We
can revisit this at a later time when it is clearer how we can reduce
user friction.  In the mean time, keeping the Core API barebones is a
good idea.