chore: try to sidestep race in use of check_if_string_present_in_trial_logs test helper (#3367)
This change updates `check_if_string_present_in_trial_logs` to follow trial logs, which makes it adhere to the master's loose attempt to delay until we think all logs should be here (this is hard to do accurately, but this at least should be better). This also loosens that delay up some more. It used to be 10s but I shortened it because it makes the CLI take too long to exit; 5s should be an OK compromise.