chore: bump docker and k8s dependencies (#3352)
This change bumps dependencies for k8s and docker as necessary (maybe a few more) to address dependabot issues. 
* [Transitive dependency on `github.com/dgrijalva/jwt-go` through old k8s version is removed.](https://github.com/determined-ai/determined/security/dependabot/master/go.sum/github.com%2Fdgrijalva%2Fjwt-go/open)
* [`github.com/containerd/containerd` is an indirect dependency, by bumping our Docker client to a much newer version, now we are at  `v1.5.8 > v1.4.12` ](https://github.com/determined-ai/determined/security/dependabot/master/go.sum/github.com%2Fcontainerd%2Fcontainerd/open)
* [same for `https://github.com/determined-ai/determined/security/dependabot/master/go.sum/github.com%2Fopencontainers%2Fimage-spec/open`, we are now `1.0.2 >= 1.0.2` by bumping docker](https://github.com/determined-ai/determined/security/dependabot/master/go.sum/github.com%2Fopencontainers%2Fimage-spec/open)

There were a few breakages to deal with; k8s started using `context.Context` everywhere in client-go, added a new API or two (I didn't upgrade any further as to not add client support for APIs clusters we are installing on may not support) and docker did some stuff too.