ci: make checkpoint gc tests actually wait for gc (#3403)
When the checkpoint gc tests were written, checkpoint gc ran directly on
the same compute node as the experiment ran, so waiting for 35 seconds
for checkpoint gc to finish was a reasonable design.

Now that we run checkpoint gc in a separate resource pool which might be
empty, we need to actually wait for the checkpoint gc to run.  There's
still a hard-coded 5-minute timeout, but cpu instances usually don't
take too long to spin up, so that should be reasonable.