chore: remove tools.go+go.mod tool dep tracking pattern, just go install versions of tools (#3383)
We used to track versions of tool dependencies in the `go.mod` files (as was pretty normal for go). We would `go get` it to pin a version in the `go.mod`, add a line to `tools.go` so it was "used" and then go install it (this is what cloudforest did, some cases were slightly different). It's really nice, though, to separate tool dependencies from code we compile into the master. This means the `go.mod` is much more straightforward when it comes to questions like 'how did this get here?', and in general there should be fewer funky deps and spurious security alerts.