chore: upgrade golangci-lint to latest (#3379)
This change upgrades `golangci-lint` to the latest version to use the `exhaustivestruct` linter on the expconf package. There was recently a bug where a mapping was incomplete and some fields from `model.Enviroment` weren't passed to `expconf.EnvironmentConfigV0`, resulting in a user being unable to use capabilities. In order to avoid other, similar bugs, I've enabled `exhaustivestruct` for every struct in the package and made sure they all conform or that it is OK when they don't.