fix: make systemd socket activation actually work (#3459)
Since the gRPC HTTP proxy functionality currently requires the master to
know what port it's listening on, working with systemd socket activation
requires us to do more than simply using the passed listener: we also
need to figure out what port it's actually listening on. This change
allows the master to do that by inspecting various parts of
`/proc/self` (Linux-specific, but of course this only comes up under
systemd anyway). Figuring out this by reading a bunch of text files is
not the most reassuring process, but it's an interface exposed directly
by the Linux kernel, so it should be robust enough.

Alternatives would include getting the same information via the Netlink
interface (but the space of libraries here doesn't look that good) or
doing some magic with the listeners to make proxy connections run
in-process and avoid the network entirely (the nicest conceptually, but
will require more fiddling/investigation).