feat: basic request audit logging [DET-6370] (#3516)
This change adds basic audit logging for master endpoints in the form of echo middleware, which defaults to off but can be enabled with a (for now) hidden configuration switch. To unhide and document will happen separately. Implemented as designed in ticket, so I will defer a detailed explanation of the intended functionality to Jira.