fix: add abort mechanism to stop actor deadlocks, check resource staleness on receipt (#3519)
This change aims to prevent a bug where an allocation may receive resources when it no longer needs them. From static analysis of the code, it should be semi-apparent to the reader that the scheduler may send the allocation actor resources as it pleases so long as it has not received notice that the resources have been released, and also that the allocation actor checks no preconditions upon receipt but moves straight along with launching on the resources. This change has 3 parts: ensure we always log allocation errors, since it was very annoying in debugging, make the allocation actor check the precondition "am i still pending" in the event it decided to exit already and the trial already received the allocationExited message, and as a last line of defense, don't make a bidirectional blocking ask forever.. timeout eventually.

(cherry picked from commit 00befce7ea7804592fb572c2f6d89d29c91cd223)