chore: Add logging - checkpoint storage validation [DET-6623]
We should make it clear that the action we are taking at the beginning
of a trial is to validate the checkpoint storage. In the event that
storage hangs, it isn't clear what the system is doing. This will
hopefully hint to folks that they need to look at their storage.