feat: make notebook timeout more configurable (#3498)
Previously, we could only set the idle timeout for notebooks to start
counting if there were no kernels; that requires users to manually shut
down kernels to take advantage of the feature. This change allows the
notebook idle timeout to be based on either the presence of connections
to kernels or the last activity on any kernel or terminal (as well as
the method from before).