chore: remove container_id from internal apis
Even though agents and k8s let us see the container ids before we
actually start the container, for other runtimes we plan to support
(like slurm) are likely not to support such a workflow.  A simpler, more
universal workflow is to ask some external resource manager, "please
give us N containers with this environment" and we get N containers with
identical environments.

This means that we should not be relying on container_id anywhere in the
harness (except perhaps Rendezvous API which can just use a different
implementation at some point when we have to support non-agent, non-k8s
rendezvous).  But the ClusterInfo API, which will become a public API
soon, should certainly not expose a container_id.