chore: unequate reservation existence and the fact they've been started (#3577)
This change adds a new bool to the allocation actor to track when reservations have been started. test_cancel_ten_experiment was flaking occasionally, and it was caused by falsely equating len(reservations) > 0 to "there are containers we should try to kill, and we should wait until we receive receipt of them being killed". This wasn't true when they are set at the top of ResourcesAllocated and we exited due to a DB failure, the experiment already being canceled, etc, before starting them. This caused allocations to hang waiting for ContainerStopped messages, which agents never send if you request a non-existent container to stop. This should make this happen less but still isn't perfect; in the event a container fails to launch this may still happen (since we would think we started it, but didn't). But, when starting reservations is truly an ask that can return all errors from the agent, this works (everything being completely async is slightly painful). This work is happening separately.